+++ src/Share.jsx
@@ -7,7 +7,6 @@ import useStore from './store/useStore'
 import useTheme from './Theme'
 import debug from './utils/debug'
 import {ColorModeContext} from './Context/ColorMode'
-import {getNewModelRealPath} from './utils/file'
 import {handleBeforeUnload} from './utils/event'
 // TODO: This isn't used.
 // If icons-material isn't imported somewhere, mui dies
@@ -82,7 +81,6 @@ export default function Share({installPrefix, appPrefix, pathPrefix}) {
             appPrefix={appPrefix}
             pathPrefix={pathPrefix}
             modelPath={modelPath}
-            getNewModelRealPath={getNewModelRealPath}
           />
         </ThemeProvider>
       </ColorModeContext.Provider>
